+++ R/birdnetR-defunct.R
@@ -1,0 +1,11 @@
+#' Defunct Functions in the birdnetR Package
+#'
+#' These functions are gone and no longer available.
+#'
+#' @name birdnetR-defunct
+#' @keywords internal
+#' @section Defunct functions:
+#' \describe{
+#'   \item{\code{\link{install_birdnet}}}{This function is defunct. Python dependencies are managed using `reticulate::py_require()`}
+#' }
+NULL